refactor: extract Wan VAE implementation
